Activate ADVANCED_OK
ADVANCED_OK provides some output for the buffer usage.

Now we get:
ok N34 P10 B1
With N = Linenumber, P = free entries in the planner buffer, B = free entrys in the buffer for preparsed g-codes.

Sadly no information about the RX-buffer.

A test with RepetierHostSettings_2.png gives Test_3.txt.

First search for "error:" - the errors seem to occure much later.
Then search for "N40" - Why do we OK that so often? Likely these are the commands we encued with G29. We have not had linenumbers for them. But G29 was sent with N26? Someting is wrong.