Extend ADVANCED_OK
by R = free places in the RX-buffer
and the output of the g.code line we just finished.

See Test_4.txt
The original ADVANCED_OK sends a N-number not matching the actual line.
The number send corosponds with the line just put __in__ the comand buffer, not the one we finished.