Repair N in ADVANCED_OK
Send the correct line number.
If no N was in the command - non't send an N.

See Test_5.txt
Now the linenumbers do make sense.

Let's have a look on the R values. Until the first G29 they are 128 - unused, everything fitted into the B-buffer.
Behind the first G29 we begin to use the R-buffer.
Somewhen behind the second G29 the R-buffer is completely full and we get errors.